refactor(fonts): unify the duplicated glyph decoder
decode_text_to_unicode existed twice — an extraction copy in
extractors/text.rs and a rendering copy in rendering/text_rasterizer.rs
— and the copies drifted in both directions. UTF-8 codespace CMaps and
the preserve_unmapped_glyphs toggle existed only in extraction;
ligature decomposition and dropped-glyph accounting only in rendering.
A Type0 font with a UTF-8 CMap extracted right and rendered garbage; a
ligature rendered right and extracted lossy. The forked neighbourhood
(fallback_char_to_unicode, get_byte_mode, TextCharIter, the TrueType
cmap donor selection) had the same disease: rendering's get_byte_mode
lacked the codespace-authoritative two-byte check and the bare H/V
CMap handling, and its cmap donor selection was first-wins and
nondeterministic where extraction picks best coverage with a
deterministic tie-break.

One decoder now lives in fonts::unicode_decode, carrying the union of
both feature sets. The two genuine policy differences stay explicit in
DecodePolicy: extraction preserves unmapped glyphs when asked and never
decomposes ligatures (that happens downstream); rendering drops
unmapped glyphs but tallies and reports them, and decomposes
presentation-form ligatures for the shaper. Extraction output is
unchanged; rendering gains the extraction-only fixes.
